fix: CORS wildcard + file_size_mb schema type + global error handler
- CORS: allow_origins=["*"] eliminates Render FRONTEND_URL misconfiguration
  (safe because auth uses JWT Bearer tokens, not cookies)
- datasource schema: file_size_mb int->float fixes Pydantic V2
  serialization error causing 500 on /api/data/sources
- strategy schema: creator_id Optional to handle NULL in old records
- Global exception handler: returns JSON error details instead of opaque
  500, and logs full traceback for debugging

+++ backend/app/schemas/datasource.py
@@ -12,7 +12,7 @@ class DataSourceResponse(BaseModel):
     date_from: str
     date_to: str
     columns: str
-    file_size_mb: int
+    file_size_mb: float
     source_type: str = "upload"
     broker_name: str = ""
     # Instrument profile
